SistemaPOs/requirements.txt
Flask>=3.0
gunicorn>=23.0
httpx>=0.27
psycopg2-binary>=2.9
openpyxl>=3.1
reportlab>=4.0
qrcode[pil]>=7.4
